Breaking: Move library packages to pkg/
