+++ pkg/encoding/display.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"strings"
 
-	errawr "github.com/puppetlabs/errawr-go"
-	"github.com/puppetlabs/errawr-go/impl"
+	"github.com/puppetlabs/errawr-go/pkg/errawr"
+	"github.com/puppetlabs/errawr-go/pkg/impl"
 )
 
 type ErrorDisplayEnvelope struct {
