+++ pkg/impl/item.go
@@ -1,7 +1,7 @@
 package impl
 
 import (
-	errawr "github.com/puppetlabs/errawr-go"
+	"github.com/puppetlabs/errawr-go/pkg/errawr"
 )
 
 type ErrorItems map[string]errawr.Error
