+++ pkg/testutil/stub.go
@@ -3,8 +3,8 @@ package testutil
 import (
 	"fmt"
 
-	errawr "github.com/puppetlabs/errawr-go"
-	"github.com/puppetlabs/errawr-go/impl"
+	"github.com/puppetlabs/errawr-go/pkg/errawr"
+	"github.com/puppetlabs/errawr-go/pkg/impl"
 )
 
 func NewStubError(code string) errawr.Error {
